add pss to procstat
- refactor procstat to support OS-specific metrics
- add pss as a Linux-specific metric to procstat
- update `procstat/README.md` correspondingly
- bump `gopsutil` version in `Gopkg.toml`
