+++ Gopkg.toml
@@ -148,7 +148,7 @@
 
 [[constraint]]
   name = "github.com/shirou/gopsutil"
-  version = "2.18.05"
+  version = "2.18.07"
 
 [[constraint]]
   name = "github.com/Shopify/sarama"
