+++ plugins/inputs/procstat/process.go
@@ -4,25 +4,8 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/shirou/gopsutil/cpu"
 	"github.com/shirou/gopsutil/process"
 )
-
-type Process interface {
-	PID() PID
-	Tags() map[string]string
-
-	IOCounters() (*process.IOCountersStat, error)
-	MemoryInfo() (*process.MemoryInfoStat, error)
-	Name() (string, error)
-	NumCtxSwitches() (*process.NumCtxSwitchesStat, error)
-	NumFDs() (int32, error)
-	NumThreads() (int32, error)
-	Percent(interval time.Duration) (float64, error)
-	Times() (*cpu.TimesStat, error)
-	RlimitUsage(bool) ([]process.RlimitStat, error)
-	Username() (string, error)
-}
 
 type PIDFinder interface {
 	PidFile(path string) ([]PID, error)
