+++ plugins/inputs/procstat/procstat_windows.go
@@ -1,0 +1,9 @@
+package procstat
+
+// Add Linux metrics into fields map
+func (p *Procstat) getOSMetrics(proc Process, prefix string) map[string]interface{} {
+
+	fields := map[string]interface{}{}
+
+	return fields
+}
